\name{fsu-package}
\alias{fsu-package}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Forecasting Functions for Time Series
}
\description{
Featuring methods and tools developed by the Foreacsting & Strategy Unit (National Technical University of Athens) \cr
\url{https://fsu.gr}.
}
\details{
Package: fsu \cr
Type: Package \cr
LazyLoad: yes
}
\author{
Evangelos M. Theodorou \cr
\cr
Maintainer: vagtheodorou@fsu.gr
}
