% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\docType{package}
\name{FSU}
\alias{FSU}
\title{The FSU Package}
\description{
This package contains handouts for
Marc Ratkovic's Short Course on Machine Learning and Causal Inference
}
\section{Unpacking Handout Materials}{


Use \code{\link{get_handout}}, with a number to unpack handout materials
in a local directory.  For example:

\code{get_handout(2)}

This will unpack the second handout materials into a folder called
'handout2` in your current working directory.
(Type \code{\link{getwd}} if you're
not sure where that is.  You can change it using \code{\link{setwd}} or
through the graphical user interface).
}

\section{Starting Again}{

If you want to start again on the same handout, you'll can unpack the
handout materials again under a different name (\code{\link{get_handout}} won't
overwrite an existing folder.)  To do this use the
\code{newname} argument.  If you want your new copy to be called
'handout2-forreal', then use

\code{get_handout(2, newname = "handout2-forreal")}

Provided there's not already a folder of that name in your
current working directory, you'll get a fresh set of handout materials
unpacked there.

If you want to preview the questions in a handout materials without
unpacking it into your local file system, use \code{\link{preview_handout}}.
}

