% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{preview_handout}
\alias{preview_handout}
\title{Preview Handout Materials}
\usage{
preview_handout(pnum)
}
\arguments{
\item{pnum}{handout number}
}
\value{
Nothing.
}
\description{
Launches the pdf viewer to show the questions in some handout materials.
}
