% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{update_package}
\alias{update_package}
\title{Update the package}
\usage{
update_package()
}
\value{
Nothing
}
\description{
This function updates the package by running
\code{devtools::install_github("ratkovic/FSU")}.
}
